pyproject.toml
[tool.poetry]
name = "previsao-de-resultados-de-partidas-de-futebol-usando-redes-neurais-de-grafos-heterogeneos"
version = "0.1.0"
description = ""
authors = ["JoaoPasseri <jvpmamorim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
numpy = "^2.1.1"
ipykernel = "^6.29.5"
duckdb = "^1.1.0"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
